Another big incomplete source-dump.
* Partially implemented the ITLB and UTLB Translation lookaside buffers
* Added skeleton of future "dumb" memory device to be used for testing
* Added BaseException.hpp from the CHIP-8 emulator I wrote last week.

I don't know if the TLB stuff and the memory caching stuff is even necessary
(the cache might be entirely transparent to Dreamcast games and the TLB might
not ever be used since a lot of game systems from that era run everything in
kernel mode with no memory protection) but for now I'm implementing it anyways
so I can make sure I understand the big picture of how everything in the SH4
fits together.  Later I might remove a lot of this code.

This still doesn't build, but when I finish the memory caching stuff and the
CPU exception support I hope to be able to build a basic "stub"/hello-world
program, and from there I can start implementing some opcodes and gradually
getting the SH4 interpreter towards basic functionality.