Move instruction-cache code into a separate file.
Next I will do the same for the operand cache code.

This necessitated refactoring the read4 function to pass the buck down to its
callr wrt checking the iix bit in the cache-control register because that
register is a private member of class sh4.  I do not anticipate this to be
a problem because sh4 is the only class that will ever need to use Icache
anyways,