Add BasicMemTestWithOix
This is a copypaste of BasicMemTest, but it also sets the OIX flag before doing
anything.

I also cleared bits 7 and 8 of the cache line entry selector in
Ocache::cache_write_back.  This is not needed to make the test pass, but it
makes me feel better.  These bits correspond to ORA and OIX respectively, and
are redundant in this case since we also have them in the tag.  I'm worred that
if OIX is enabled, then the OIX bit would have to be remapped from bit 13 to
bit 25 somehow.