unit_tests: template-ize sh4_test.cpp
This will make it easier for me to add new testcases in the next commit.
Also, the part where it checks the instruction cache in BasicMemTest::run
no longer checks both halves of the 32-bit int (since that would be hard to do
because it might not be reading a 32-bit int or even an in which is larger than
inst_t).  However, I initially forgot to implement the part where it actually
checks the value of the inst_t against the value it read in, so it's actually
testing more now than it was before this commit.