Add all the boost deps that are in the top-level boost directory
The reason for this is that I want to in the future be able to easily add new
boost sub-libraries by adding the directory named after the library without
struggling to figure out which top-level dependencies I may also need.

Adding all of boost is not an option cause it's really huge.