Simple unit test-case for UTLB.   This test maps a single page into memory and then tests that all memory writes which fall outside that page generate the proper CPU exception (and also that all memory writes within that page go through).
I also fixed a bug caused by a typo in the code which raises the CPU exceptions.

Obviously there are a lot more unit tests for the MMU that need to be written
to have good test coverage in this area.

One more thing I'm concerned about is memory access that spans multiple pages, I
don't think I've handled this corner case yet and I definitely don't have any
ptests which can handle it.  The solution might be to make this ptest read back
the memory and verify that it matches expectations.

Surprisingly enough, the only bug I've found so far was a misprint in the code
that generates the exceptions.