sh4asm: (partially complete) assembler library for sh4
This will be used in the unit testing code to test the CPU interpreter when that
is ready.  It may also be useful for other debugging purposes.

So far all I have are opcode tables and some (incomplete) code for handling the
tokenization.  IDK how I'm going to do the actual parsing yet.

The opcode tables so far cover all opcodes up to (but not including) the
floating point ones

As I wrote this, it became increasingly clear that I should be employing some
sort of templated/metaprogramming approach.  I am not sure how that will work,
but this commit is a massive mess and I do not believe I am currently
implementing the sh4asm in a clear, readable way.