Wrote basic skeleton of how the sh4asm will actually work.
The stuff I did last night is trash, I'm holding onto it for now because I can
t save the time it took to type most of that in by copying/pasting it from the
old implementation instead of typing it back in again.

The new implementation is template based, and it links the decoding together
with the assembly quite nicely.