sh4asm: implement .byte directive to insert raw data into the program
Also implement a binary mode to be used as an alternative to hex strings so you
can disassemble the bios and assemble files that actually run in the emulator

This resulted in a lot of sh4asm's code getting refactored.  I came up with a
new paradigm wherein the Sh4Prog has a binary copy of the programm and a
text-asm copy of the file that are both updated in sync with each other.  This
seemed like a really clean way to do things, but then I ended up with 5-minute
assembler/disassembler times because it has to keep resizing std::vectors, so
I will probably end up going back to the old paradigm eventually in the future.