sh4 interpreter: initialize the pc to 0xa0000000
I figured out that the reason this works is that the top 3 bits get cleared
when it tries to access the actual memory value, so it is mostly equivalent
to setting pc=0.  So now I know that I'm *not* mapping the bios in to the wrong
place like I thought earlier

Also make the GdbStub respond to any improper memory accesses with E01.