sh4 interpreter: fix a number of MMU-related bugs
Prior to the great memory remapping of aea12c500d63183463400f9bb56e65d8fc341bc7
the mmu tests in sh4mem_test were mapping 0 (vaddr) to 0 (paddr); this was
inadvertantly covering up several mistakes.  Now that 0 is unwritable, the test
has to map 0 (vaddr) to something else and all thos bugs have to be fixed.

And thus they are.