sh4asm: introduce new buffer-free API that lets programs not use Sh4Prog
The reason for this is that Sh4Prog's excessive buffering was causing a lot of
slowdown when I was assembling/disassembling the Dreamcast's BIOS.  Sh4Prog is
now built on top of the new interface.

Unfortunately it's still slow, but at least it's not as slow.  TBH this
probably wasn't worth the effort I put into it since it's just sh4asm and not
part of the emulator proper.