minor build fixes
One was in sh4mem_test, where g++ was concerned about some variables it thought
might be used without being initialized.  They weren't being used
unitinitialized because the ORA flag isn't changed during test execution, but
the compiler has no way of knowing that so I initialized them both to 0.  This
was only happening with -DCMAKE_BUILD_TYPE=Release.

The other problem fixed was where clang++ was complaining about some
comparisons being done in MemoryMap.cpp which were effectively NOPS (always
true).  I had done this deliberately for the sake of code clarity, but I took
them out because this is the price you pay for building with -Wall -Werror.  I
don't know if this was exclusive to -DCMAKE_BUILD_TYPE=Release because I don't
test on clang++ very often and I don't care enough to try it out with
-DCMAKE_BUILD_TYPE=Debug.