sh4 interpreter: fix several places where instructions were modifying sh4 state on failure
This also necessitated a change to the unit_tests for 0010nnnnmmmm0100,
0010nnnnmmmm0101 and 0010nnnnmmmm0110.  After consulting the sh4 software
manual, I realized that these instructions are supposed to write what the
source value was *before* the instruction.  Previously, these opcodes would
write what the source value was after the decrement of the destination address
register; this was leading to incorrect behavior when the source register and
the destination address register were the same register (since the value written
would be decrimented).