sh4 interpreter: implement all the opcodes which touch FPUL
There's some error handling described in the sh4 spect which is not implemented
because it would be really slow and also difficult to implement...maybe I
should reconsider this in the future.

Also I had to move to C++11 so I could get fesetround and isfinite.
I don't like C++11 very much.