sh4 interpreter: Don't handle delay slots within branch instructions
This way, I can step through delay slots in gdb instead of stepping over them

As a result of the way this commit is implemented, we no longer raise an
exception when software attempts to execute a branch/jump instruction in a
delay slot if the delayed-branch instruction that triggered the delay slot
evaluated to false (did not branch).  This is an innacuracy, but it's a
relatively minor one which avoids having to do an extra branch in the
interpreter.  I may want to reconsider this in the future and opt for the more
accurate behavior.