Make travis-ci use clang++ instead of g++
Travis uses an older version of g++ which doesn't seem to have C++11 support.
I'm hoping that their clang++ version is more up to date.

This is another great reason not to be using C++11.  The only reason I turned on
C++11 in commit 82a9446b57fd11e542c5733c82d6d8d0602dbb07 is that I need it to
get access to certain C99 features which are inaccessible from C++03 because
C++03 is based on an older version of the C standard.  Those features are
isfinite and fegetround/fesetround.  I am currently considering alternative ways
to get these features, such as implementing them myself.