sh4 interpreter properly tick the tmu based on instruction execution.
This makes the assumption that the execution time of each instruction is equal
to its issue cycle count because we assume perfect cache (even though there *is*
Ocache/Icache emulation code, it doesn't effect the timers), and no pipeline
stalls.  This does not take the dual-issue nature of the pipeline into account.