sh4 interpreter: implement DMAC registers, but don't actually implement the DMAC
We log reads/writes to the DMAC registers to std::cerr for now.
So far the PC only makes it far enough to initialize the DMAC, and I'd rather
wait until I can get a better idea of how it's used in Dreamcast software
before I actually implement DMA transfers.