sh4 interpreter: implement the interrupt controller
One side-effect of this commit is that the sh4tmu_test now reports the value
of R3 as 2 instead of 1.  I believe this is because of the way the extra
instructions added to the test's sh4 program have offset the value of the sh4's
cycle_stamp.  Specifically, the first time the "DT R3" instruction gets run,
the value of "print/x sh4->cycle_stamp - sh4->tmu.last_tick" is now 0; it used
to be 2.