refactor the g1 bus mmio regs to use the same format as the maple and sys mmio regs
One difference is that for length-checking, I allow reads/writes that are less
than or equal to the length of the register (as opposed to the other mem-mapped
regs, where I only allow reads/writes equal to the register length).  This is
because the bios tries to do a 2-byte write to the 4-byte G1RRC register.  In
the future I may want to also allow writes less than the register length for
all memory-mapped registers; for now I only allow it in cases where it is
actually needed.