link against librt
Travis-ci is getting build failures due to clock_gettime not being available.
I believe this will fix it (although for some reason everything works perfectly
fine on my PC with no need for this library).

Also make a change to .travis.yml that will hopefully cause build failures like
this to actually be reported as failures if they ever happen again.  This
failure has evidently been happening for almost two weeks now, but this is the
first I've heard of it because travis was always logging that the build passed.