shut up warnings for SB_FFST and CHCR3 read/writes
I'm doing this because the warnings cause excessive log spam when booting
kallistios programs; the log spam significantly increases boot time.

Usually I like log spam when guest programs try to use features that haven't
been fully implemented because it tells me what I need to go implement, but in
this case it's impacting functionality and getting in the way.