Debugger: implement softbreaks
This also fixes an issue with the M (memory write) GDB packets not working.

This also *sorta* implements TRAPA, but only when the debugger is enabled

This also implements the general illegal instruction exception.  I made the
somewhat odd behavior to turn this into a softbreak when the debugger is
running.  This may be unnecessary since it looks like gdb always uses a TRAPA
for softbreaks, but whatever.