Debugger: increase the number of watchpoints, rwatchpoints and hard breaks to 16
Earlier it was 10, and there have been a couple occasions where I almost hit
that limit.  It also seemed weird since 10 is not a power of two.

There is a performance tradeoff here since it has to check that many more
addresses every time it does something, but only if the debugger is enabled.