ape the GD-ROM startup sequence
This commit doesn't implement the dreamcast's gdrom drive, but it does do
everything it needs to do to make KallistiOS think it's initialized.

I don't like to make messy commits like this, but I've spent all weekend
figuring this out and it's driving me insane and I don't want to make any
changes lest I jinx it.  I'll clean this up later.

This also fixes a bug with a bunch of the registers maps where they weren't
handling 4-byte registers properly.  Ideally this part should be a separate commit,
but once again I just need to get this committed while it works.