AICA: Add workaround to get the RTC sort-of working
I hardcoded it to approximately Jan 1, 2000 at 12:00 AM.  It never
increments because I'd like to hold off on implementing AICA for real
until after I've written a real scheduler for the CPU.

The reason I'm doing this is so that I can play the BSD cave adventure game
from KallistiOS' examples directory, and it translates time to it's own epoch
which is based at 1977.

If the RTC is set prior to that, it'll break the game's logic and refuse to
start.  The reason for this is that there's a snarky joke where it tells you
that you need to be more patient if you save your game and then immediately
load that saved game, and the logic for implementing that will return a false
positive if the current time is negative.