sh4 interpreter: don't use register write handlers to initialize registers
This only applies to the SH4 registers, not any of the other components.

This is needed because it was fucking with the SCIF and causing it to send
garbage data during initialization because the write handlers thought the
software was using it to transmit data over serial.  In general, using the
write handlers for initialization was a bad idea due to registers like this
that have side-effects.

This also lets me ditch that lame squelch_warnings hack from commit
d82746905119de7c1d18d4d0da90ef1329394075