Handle SIGINT by setting is_running = false
This way, it prints the performance stats before exiting when you press CTRL+C.
Previously, the only ways to do that were to quit from within gdb (which throws
off the performance stats big-time since you have to interrupt execution) or to
do something that would make the interpreter throw an exception.

This commit also makes the emulator print the reason why it ended execution.