unit_tests: increase remaining loops max for sh4tmu_test to 3
This is a response to a test failure introduced by 9e66e8c8a0a5e874c7bf372d68554694f6fc833d

The problem is that the test is now exiting with 3 remaining loops instead of 2.
The reason this happens is that the TMU countdown starts 1 sh4 cycle earlier because the
TMU countdown is rounded to the sh4 clock cycle corresponding to the beginning of the
previous TMU clock cycle.

I'm not sure if maybe the value should be 4 (or maybe even 6 since there are 4 sh4 cycles
for every TMU cycle) because I haven't fully thought this through, but I'll cross that
bridge when I get to it.  For now, I feel fine leaving it at 3 because less is always better
and it *does* make sense that the number of remaining loops would goe up by at least one due
to the changes introduced in commit 9e66e8c8a0a5e874c7bf372d68554694f6fc833d.