Turn off travis-ci.
Managing dependencies with travis-ci is too difficult because their system
tends to run off of really out-dated versions of Ubuntu Linux with really
outdated package repos and in general it's primarily geared towards the webdev
crowd.  I do appreciate their service and some day I intend to use it again,
but for now it's creating more work than it's saving.