sh4 interpreter: refactor FPU opcodes to check for the PR/SZ bits themselves
This is a precursor to the big O(1) decoder upgrade.  With this commit in place,
I shouldn't have to worry about instructions with overlapping opcodes (which
the FPU has quite a lot of) breaking when the O(1) decoder is implemented.

I tested this as much as I could, however a lot of the FPU opcodes are still
unimplemented, so I couldn't test the SZ/PR bit-checking code for their cases.

Also, the programs in my testing set don't use the FPU much in general.