get rid of the libc code I had previously included to give C++03 some of C99's fenv functionality
Now that I'm on C11, I don't need this.

I also moved the C++ code in sh4asm and unit_tests over to C++11 so that they
can use the one true fenv, too.  Generally speaking I don't like C++11, but as long as
it's not in the emulation core idgaf.