GD-ROM: handle 0x70 packets "correctly"
Nothing much changes other then the fact that I set the int_reason register
correctly, although that doesn't even matter since the bios was never checking
it to begin with.

Based on register access logs I'm pretty sure that the BIOS doesn't expect to
receive any data from this command.  I'd love to verify that in the machine
code, but the bios GD-ROM functions use a sort of co-operative multithreading
that makes disassembly pretty unwieldy.

Based on what I've seen in the Linux Kernel's GD-ROM driver, this packet has
something to do with spinning up the disk, so it makes sense that software
would not require a response.  I'd still love to verify this, though...