compilation of weird things I had to do to get this building on my laptop
The laptop runs Debian and my desktop (where most development work is done)
runs Gentoo.  I can't explain why this was working on the one but not the other.

This seems to be working on both the desktop and the laptop (SO FAR).
I'm sure things are going to get real fun later when I add in FreeBSD support...