bare skeleton of framebuffer sync code
The purpose of this code is to update the framebuffer in PVR2 texture memory
from the opengl target (ie the fbo) when they're out of sync.  Then, for the
output stage, we turn that framebuffer into an OpenGL texture and render it for
real as a quad stretched across the entire screen.

Currently it doesn't actually load the OpenGL view of the framebuffer into the
PVR2 view of the framebuffer because I haven't written the conversion functions
yet, but the code that decides when to call them is there.