Implement more of the rendering pipeline
At this point, we have all the basic stuff except the part where we turn
triangles from the TA into pixels.  that will happen in a later commit.

This commit displays a single hard-coded triangle because it is concerned
with setting up the system for syncing the OpenGL frame buffer to the PVR2
in-memory framebuffer.