fix two deadlock cases in the rendering code
One is the possibility that the emulation thread is waiting for the renderer
to render a frame after it has already exited.

The other is the possibility that the renderer will need to draw more than one
frames to catch up to the emulation thread.  I chose to make it draw all
pending frames; arguably it might be better to draw the latest frame and skip
the older ones, but I don't want to introduce that into my emulator just yet.