New shutdown protocol
if the gfx_thread thinks it's time to exit, it tells the main thread to close
and then the main thread tells the gfx_thread to exit.  Having a single point
of exit should fix those occasional deadlocks I was seeing when exiting the
emulator via the window manager's X button.