gdb_stub: clean up the way we handle register writes
The purpose of this commit is to make the bank-switching behavior more
predictable from the perspective of a user on a remote GDB client.

To this end, the following changes have been made:
* writes to r0b0 through r7b0 and r0b1 through r7b1 are no longer accepted.
  This removes the ambiguity that can arise when a G packet comes in where
  the new values of r0-r7 do not match the new values of the banked versions.
  Users can still edit the banked versions of r0-r7 by manually toggling the RB
  bit in SR
* the P packet only touches the one register that actually changed.  This way,
  we don't accidentally write to r0-r7 when the RB bit is being toggled.
* For the sake of G packets, the SR register is always evaluated first so that
  the new value of RB is what determines the bank that the writes to r0-r7
  effect.

These new policies have not been rolled out to the FPU registers because I
don't have bank-switching implemented there yet anyways.