sh4 interpreter: add INVARIANTS-only tests to make sure proper instruction callbacks are called
This has actually unmasked a bug (which I haven't fixed yet).

This is mostly just a problem for the FPU instructions (since several of them
overlap with one another), but the sake of posterity I did implement this
for the rest of the sh4 instructions.

In theory I could use all the preprocessor macros I made here to replace
sh4_compile_instruction, but for now I'll keep them both because I like having
the two redundant systems in place because I can use them to validate each other.

Clean up the last two commits a bit