PVR2: basic texture mapping implementation
This version is messy and deficient and needs more work, but I have to get onto
a plane later tonight so I'm committing it as-is.

For one thing, non-twiddled textures won't work because it uncondtionally
de-twiddles all textures.
For another thing, it doesn't allow more than one texture to be used at once
because there's no infra in geo_buf to switch polygon contexts mid-render yet.

Also I'm not 100% sure if I remembered to make it update the textures when the
user over-writes them in texture-memory.  There is definitely infra in the
texture-caching code to support that, but I can't recall if I went all the way
or not.

Also there's no code to evict old, unused textures from the cache aftetr a while,
that needs to happen.