don't use GL_ARB_explicit_uniform_location for fragment shaders
This causes segfaults on the intel i965 my laptop uses.  My desktop
(which runs twin nvidia GTX 980s in SLI) has no problems with this.

I did verify that this bug is not caused by any memory corruption within
WashingtonDC by reproducing it in a smaller program I wrote that does nothing
but load a vert shader and fragment shader and link them together.  Because of
this, I'm inclined to believe that I have stumbled upon a driver bug.

Even if it was incorrect to use explicit uniform locations in fragment shaders,
I would expect glLinkProgram to fail instead of segfaulting.

For consisteny's sake, I may take the vertex shaders off of
GL_ARB_explicit_uniform_location in the future, but for now they stay.  I have
not experienced any problems from using this extension on the vertex shaders on
any machine I own.