PVR2: pass the z-values along to OpenGL for rendering
This necessitates storing the minimum/maximum z-values of each geo_buf and
dividing by those to get the opengl clip coordinates because PVR2 does not
seem to have any concept of a clipping plane.  Thankfully, PVR2 also does not
seem to have any concept of a persistent depth buffer, so it's alright if the
z-coordinates get scaled differently every frame as long as they're scaled
consistently within each frame.

hidden-surface-removal/depth-sorting is still not implemented, so the last
polygon to be rendered will still always be on top no matter what.