PVR2: call finish_poly_group whenever an end-of-list packet is received
this should hopefully resolve the problem with the invalid blend factor indices.
If I am correct, then the root-cause is that not all of the fields in the last
poly_group structure were being initialized because finish_poly_group was not
being called on it.  This would cause the blend factor indices to sometimes be
out-of-bounds.