GD-ROM: do all logging through a common macro
the idea behind this is that when I'm pouring through logs to debug GD-ROM
problems, I can sort the GD-ROM messages out from the rest of the logs by
grepping for "GD-ROM".

Eventually I'd like to have a common logging infrastructure which every
subsystem in WashingtonDC uses, but for now I'm doing it only in certain areas.