Make the gfx_thread print pending events on exit
This is instrumentation to help me debug github issue #12.  The idea is that
next time this occurs, I will be able to kill the program (by CAREFULLY
manuevering the mouse around the edges of the window to avoid waking it up as I
click the X button) and observe if there are any events pending that the
gfx_thread has not yet serviced.