sh4: don't allow unprivileged mode
Earlier I did allow unprivileged mode, but I would raise an ERROR_UNIMPLEMENTED
if a guest program attempted to access memory in SH4_AREA_P0 while unprivileged.

This (in theory) reduces compatibility, but it also gets rid of an expensive
check that was happenening every time the user tries to access memory so it's
worth it for now.