gfx_thread: get rid of fb_out_lock
There was a deadlock caused by a lock-order reversal between gfx_main and
gfx_thread_read_framebuffer.  I could have fixed the reversal, but it occurred
to me that there's no reason for these two separate locks to exist so I got rid
of one of them instead.

Now that I think of it, this code's so serialized there's not much point to
having a separate thread for graphics...