put all the code that runs in gfx_thread into its own directory under src/gfx
The interface code (ie functions that obtain information from the gfx_thread
but run outside of it) currently resides under src/gfx as well.  This will
probably always be true, but I would like to further segregate it from the real
gfx_thread in the future by putting it in its own file or something like that.