move all the code that has to touch the sh4 out of gdb_stub and into debugger
Soon I'll be moving the gdb stub into the new io_thread, and when that happens
I'm going to keep debugger.c running in the emulation thread.  Because it's not
going to be in the emulation thread, I need the gdb stub to not touch the
hardware directly.  debugger will touch the hardware for the gdb stub since it's
in the emulation thread.

This also will make it easier to implement stubs for other debuggers if I ever
want to do that for some reason.