io_thread: make io_thread_kick activate a user-triggered event
The old way was to call event_loopbreak and do the event processing from
outside of eventlib.  It has occurred to me that this caused a race condition
because event_loopbreak is a no-op if the io_thread is not blocking on
event_base_loop; thus it was possible for io_thread_kick calls to get lost if
they happened while the io_thread was already awake.