sh4 interpreter: fix the way sh4_single_step handles the dual-issue pipeline
The new SH4_GROUP_NONE value for last_inst_type is used to indicate situations
where the second pipeline doesn't issue because it issued the previous
instruction.

Also I split up the sh4_single_step function a little so that the
fetching/cycle calculation is done in a separate helper function to make things
simpler.