move the gdb_stub into the io_thread
the debugger code itself remains in the emulation thread.

One thing I'm not so proud of is the deferred command queue in the debugger
code.  The purpose of this is to make sure that all of the code in the debugger
that touches the hardware does so from the emulation thread because I would
rather do that than add locking into the emulation thread.  I implemented this
with a FIFO protected by a mutex.  It looks over-engineered and messy, but it
works and I feel confident that it won't fuck up if the remote debugger rapid-fires
commands faster than I can fulfill them (this is why i used a fifo instead of a
single buffer).  In the future I may move this into the gdb_stub.

Also the emulation thread spins when it is waiting on the debugger; I need to
find a way to put it to sleep or something.