add a CLI switch to enable/disable wireframe rendering
This will hopefully be helpful for debugging the z-fighting problem because
I can use this to get a good look at what's being rendered.

Currently it doesn't disable colors and it looks like I failed to disable
textures somehow, too.  I intend to address this soon, but for now everything
works.