initial implementation of optional system-call tracking for GD-ROM.
The point of this is to log system calls so I can debug what they're returning
to higher-level software.

The initial goal of this was to use it to debug some early-boot problems with
the firmware's GD-ROM access, but I now know that it isn't going to be useful
for that because the firmware does not use system calls until it gets to the RTC
reset screen.