Add a new hack to fool Power Stone into thinking I have working audio hardware
It still doesn't boot because I have to implement a new GD-ROM command yet.

Initially I was hoping to do something similar to this for Namco Museum, but its
initial audio access is more complicated.  I'm hoping that this will be all I
need to do to get Power Stone to at least boot to the Capcom logo.