PVR2: further simplify tex_twiddle by tying the x and y shifts together
All calls to tex_twiddle now start at shift=10 instead of the actual dimensions
of the image.  This is safe to do even for smaller images because any given
image is just the upper-left corner of a larger image, and being in the
upper-left corner doesn't set any bits.  This also eliminates the awkwardness
the original algorithm had concering rectangular textures (that awkwardness
being that one dimension would reach the zero-level before the other one).