scheduler/interpreter: correctly handle cases where new events are added to the beginning of the event list
This fixes the problem wherein Crazy Taxi was hitting an INVARIANTS assert in
the SPG code which was caused by executing an SPG interrupt late.

It works by never advancing the cycle count beyond the next scheduled event
even if it still ends up executing an extra instruction.  I believe allowing the
CPU to absorb any extra cycles is the most accurate way to handle these sorts of
discrepancies since it's very difficult for a program to observe minor
fluctuations in CPU performance.

Also improve the cycle-counting a little when the debugger is enabled by making
the single-step and non-single-step paths consistent.

This also causes the cycle-stamp to advance after every instruction executes
instead of before.  This seems like it should be more accurate, although I can't
imagine that it will cause any noticeable behavior changes.