sh4: fix inaccurate cycle counting involving the MT instruction group
Because this causes more instructions to be "free" (due to the second pipeline),
there are now more instructions that don't advance the cycle-count.  The upshot
of this is that this commit will decrease performance because on average the sh4
interpreter now executes more instructions per cycle than it used to and thus it
takes longer to run every cycle (on average).  This is similar to what the
effects would be if I raised the SH4's clock-speed.