Increase the scheduler clock frequency to 5.4 GHz
The purpose of this is to make it possible for the SPG and SH4 clocks to evenly
divide from the scheduler clock.  Currently the SPG clock is still 1/7 the SH4
clock so this commit should cause no change in behavior apart from decreasing
the amount of time that WashingtonDC can run before the dc_cycle_stamp rolls
over (which is still more than a century).

Later I'm going to set the correct clock frequency for the SPG VCLK so I can
finally rest at ease knowing that the SPG and the TMU are not giving
inconsistent timings.