Get rid of the win_thread and move all the windowing code into the emulation thread
this is needed because glfw's controller API doesn't have any sort of async
callback so I need to poll the controller state, and I thought doing that once
per vblank would be best.  TBH I'm not sure if the way I was doing it before was
portable, anyways.

Since I only check up on the windowing events once per vblank (which is
typically 60 frames per emulated second), the impact this commit has on
performance is negligible.