basic controller support
Bindings still can't be reconfigured, and the ones that are hardcoded in there
are based off the Logitech F510.  Keybinding will come eventually, but it's not
a priority for now.

This also brings analog support, so it's now possible to drive around and pick
up passengers in Crazy Taxi.

Keyboard controls are still in there with the same keybindings as before, albeit
still no analog controls support.