write a new binary emitter for sh4asm
I'm doing this to get a better grip on how compilers work since I'm going to be
writing one soon.  Also this gives me a chance to get rid of the disgusting C++
template stuff in the current version of sh4asm that keeps me from porting the
unit_tests to C (which would allow them to include stdatomic.h so I could build
them with the current codebase and start using them again).

Currently, this emitter is completely untested.