sh4asm: sh4 assembler (text) emitter
This is an emitter that outputs assembly code.  It is intended primarily for
sh4asm, although I have strived to make the code as generic and reusable as
possible.

I used a different coding style to implement this than I used on the binary sh4
emitter.  I like this better, so I just might go back and port the binary
emitter to use a coding style similar to the one used here in this commit.

This emitter has not yet been tested, so there may be some bugs.