correctly implement offset-color support
This resolves issue #22 on github (in which parts of the game timer on Crazy
Taxi were white instead of yellow) and also fixes a few other things that were
the wrong color.

Also, the text in the menu on Crazy Taxi has become visible due to this,
although it's not as dark as it should be.