Get rid of that weird half-baked ringbuffer the geo_buf infrastructure was using
In addition to being unnecessarily complicated, it probably wasn't thread-safe
since it relied on some volatile variables that probably should have been
atomic, anyways.

Ultimately, using a ringbuffer to submit data to the gfx_thread is pointless
anyways because you still need to grab the mutex to wake it up.