Minor framebuffer cleanups
Most importantly, I fixed an issue wherein the framebuffer was being fetched
from the gfx_thread twice instead of once.

Otherwise, the weird "roundabout" system where the framebuffer has to go from
the gfx_thread (which gets it from OpenGL) to the emu_thread and then back to
the gfx_thread is still in-place.  It's not impossible to make it keep the
framebuffer on the graphics server when it's not being updated directly by
Dreamcast software and I do want to make that happen eventually.