restore sh4div_test using the new sh4 emitter
The main purpose of this is so that I now have something using the sh4 emitter.
Eventually I want to turn this test (as well as sh4inst_test and maybe even
sh4tmu_test) into a standalone homebrew program that I can run on a real
Dreamcast, on WashingtonDC, or on any other Dreamcast emulator.