GFX: expand the depth range (clip_min/clip_max) slightly
Otherwise, the most-extreme edges of polygons (ie the ones that actually set
the value of clip_min/clip_max) can get culled out if the depth compare function
fails equal depth-values.  This is especially problematic for polygons which
are oriented orthogonally to the viewer because it can lead to the entire
polygon failing the depth test when it should have passed.