Use the new sh4 binary emitter to resurrect sh4tmu_test
I had to double the number of loop iterations to get it to pass;
I *think* this is because of cycle-counting improvements that have
occurred in the time since I disabled this test.  Specifically, I believe the
two instructions in the loop are now executing in parallel when they were not
before.

Also, this commit fixes a misprint in the binary emitter