remove all trace of sh4mem_test.
This was initially written waaaaayyy back in the primordial days of WashingtonDC
because I initially implemented the memory code with partially-complete support
for cache emulation and MMU (because I wasn't sure if I would need them).  That
code got pulled out a long time ago because it was deemed to be more trouble
than it's worth, so there's no need for this test anymore.