sh4asm_neo: implement disassembler
I'm using a series of switch statements instead of the lookup-table that's used
in WashingtonDC.  My theory is that using the switch statements will turn out
to be faster even though it branches because there will be less oppurtunities
for cache-misses.  I'm using sh4asm here to test out the switch approach before
I port it over to WashingtonDC-proper.