sh4asm_neo: change displacement/immediate format to match sh4asm
this is needed so that the output will be exactly the same as in sh4asm (other
than the fact that sh4asm_neo is lowercase and sh4asm is uppercase).  This in
turn is needed so that sh4asm_test can be modified to test that sh4asm and
sh4asm_neo both output the same disassembly when given the same inputs.