implement a simple, minimalistic logging infra
The main motivation for this is to provide a way to filter out log messages
at compile-time based on what class of message they are.  I'm using this to
determine what sort of an impact logging is having on performance.  The result
appears to be that logging has very little impact on performance, so I will
leave it all enabled for now.