add the new dctest version of sh4div_test as a regression test
It works by running a Dreamcast program through WashingtonDC, and comparing the
output to a file containing a capture of the same program's output taken from a
real Dreamcast.  Since both the Dreamcast and WashingtonDC were running the same
program, the output should be identical.  Certain lines which cannot reasonably
be expected to be the same (such as the lines KallistiOS prints out when it's
enumerating the maple devices plugged into the Dreamcast) are ignored.

This approach is superior to the old unit testing approach for a number of
reasons:
* It allows me to examine discrepancies between WashingtonDC and a real Dreamcast
* I do not need to come up with a specific verification for test results (a
  significant portion of the old sh4inst_test was devoted to double-checking
  operations performed by WashingtonDC).
* I do not have to maintain a special API for tests; in the past this proved to
  be more trouble than it's worth, which is why I had to drop sh4inst_test.
  Such bitrot will no longer happen with the new regression testing scheme as
  long as the cli and the serial server remain operational and keep a consistent
  interface.  Even if I am forced to change the interface for those two
  components in a way that breaks compatibility, only the evaluation script
  needs to be refactored rather than the entire test.

The corollary of the third point above is that these tests are effectively
immortal since they're real Dreamcast programs; WashingtonDC will never stop
being a Dreamcast emulator, so I will never lose these tests like I did with
sh4inst_test.

the test script is written in perl, and it gets the binary test program and
the Dreamcast capture log from external locations which are specified with the
SH4DIV_TEST_BIN environment variable and the SH4DIV_TEST_CAPTURE environment
variable, respectively.  The reason for this is that I don't want to make
KallistiOS one of WashingtonDC's dependencies, there's no need to keep
rebuilding test roms anyways since they're just data.  Additionally, there's no
way to programmatically get a new capture from a physical Dreamcast.

Eventually I intend to arrive at a single "canonical" version of every
regression test I write, but I'm not at that point yet.