sh4: fast-track instruction reads by bypassing the memory map
This boosts performance from ~8% to ~9.7%.  I believe the main reason for this
is that the instruction memory reads aren't using memory_read anymore; since
that function doesn't have a hardcoded length, the memcpy has to be a loop and
therefore it ends up causing a missed branch prediction on every call.  I should
look into getting rid of this function and replacing it with several hardcoded
length versions instead.